refactor(player): mount the engine video stage from the presentation layer
`PlayerEngine` was an application service that returned Widgets:
`buildVideoStage` mixed media_kit `Video` layout kicks and the YouTube
WebView/poster overlay tree into a transport interface (issue #664).

The stage now lives in the presentation layer. `PlayerSurfaceHost` maps
engine identity to a per-engine stage through `buildPlayerVideoStage`
(`player_stage_resolver.dart`) — `MediaKitVideoStage` /
`YoutubeVideoStage` are the relocated widgets, verbatim. ADR-0057
remounting is untouched: the host still keys the whole slot with
`ObjectKey(engine)`, so a swap (bumpRev) drops the old stage before the
old engine tears down, and `keepSurfaceWhenParked` still decides whether
a parked surface stays mounted.

The engines expose the non-widget inputs the stages read instead of
building widgets:

- `MediaKitPlayerEngine.nativeBackendAllowed` (issue #658 gate) +
  the existing `videoController`.
- `YoutubePlayerEngine.session` / `.webViewLifecycle`, plus
  `noteStageViewportSize` so the park/refocus policy stays with the
  transport code (the stage no longer logs or refocuses).

Tests: the host test injects a stage builder instead of overriding the
engine; the media_kit gate test now pumps the stage and asserts no `Video`
mounts before `prepareNativeBackend`; the poster-gating and Linux
availability tests construct the relocated stage.
