+++ docs/architecture.md
@@ -103,7 +103,7 @@ Media persists in two sibling tables (`videos` / `audios`), but callers think in
 
 ### Player surface host (ADR-0057)
 
-[`RootShell`](../lib/features/player/presentation/root_shell.dart) mounts one permanent [`PlayerSurfaceHost`](../lib/features/player/presentation/widgets/player_surface_host.dart) — the sole owner of `buildVideoStage()` for the active engine, keyed by engine identity so an engine swap disposes the old surface before mounting the new one. Viewports (vocabulary review clip, expanded player, loading stage) register [`PlayerSurfaceTarget`](../lib/features/player/presentation/widgets/player_surface_target.dart) geometry via [`playerSurfaceRegistryProvider`](../lib/features/player/application/player_surface_registry.dart) after layout and paint (so ancestor transforms are followed even when size is unchanged); the host moves one permanent `Positioned` stage to the global target bounds. When no target is attached — `RootShell` force-park for `/youtube/login`, or [`PlayerSurfaceOverlayCoordinator`](../lib/core/player/player_surface_overlay_coordinator.dart) tokens for dialogs/sheets/snackbars ([ADR-0066](decisions/0066-park-player-surface-for-overlays.md)) — YouTube stays mounted off-corner and MediaKit unmounts so the Android Surface is not first laid out off-screen (no follower transforms, which keeps WebView2 bounds correct under Windows display scaling). The first local/URL open installs MediaKit and bumps [`playerEngineRevProvider`](../lib/features/player/application/player_engine_rev.dart) so the host mounts `Video` before decode starts. Open-in-player is a single `context.replace` with a typed [`PlayerLaunchRequest`](../lib/features/player/domain/player_launch_request.dart). Feature code must not build `media_kit` `Video` / `InAppWebView` surfaces outside the host — see [ADR-0057](decisions/0057-permanent-player-surface-host.md). Enjoy dialogs/sheets ([`enjoy_modal.dart`](../lib/core/theme/widgets/enjoy_modal.dart)) default to the **root** navigator so they paint above the host ([ADR-0065](decisions/0065-enjoy-modals-root-navigator.md)); in-stage chrome uses host `overlayBuilder` instead.
+[`RootShell`](../lib/features/player/presentation/root_shell.dart) mounts one permanent [`PlayerSurfaceHost`](../lib/features/player/presentation/widgets/player_surface_host.dart) — the sole owner of the active engine's video stage (mapped from engine identity by [`buildPlayerVideoStage`](../lib/features/player/presentation/widgets/player_stage_resolver.dart), issue #664), keyed by engine identity so an engine swap disposes the old surface before mounting the new one. Viewports (vocabulary review clip, expanded player, loading stage) register [`PlayerSurfaceTarget`](../lib/features/player/presentation/widgets/player_surface_target.dart) geometry via [`playerSurfaceRegistryProvider`](../lib/features/player/application/player_surface_registry.dart) after layout and paint (so ancestor transforms are followed even when size is unchanged); the host moves one permanent `Positioned` stage to the global target bounds. When no target is attached — `RootShell` force-park for `/youtube/login`, or [`PlayerSurfaceOverlayCoordinator`](../lib/core/player/player_surface_overlay_coordinator.dart) tokens for dialogs/sheets/snackbars ([ADR-0066](decisions/0066-park-player-surface-for-overlays.md)) — YouTube stays mounted off-corner and MediaKit unmounts so the Android Surface is not first laid out off-screen (no follower transforms, which keeps WebView2 bounds correct under Windows display scaling). The first local/URL open installs MediaKit and bumps [`playerEngineRevProvider`](../lib/features/player/application/player_engine_rev.dart) so the host mounts `Video` before decode starts. Open-in-player is a single `context.replace` with a typed [`PlayerLaunchRequest`](../lib/features/player/domain/player_launch_request.dart). Feature code must not build `media_kit` `Video` / `InAppWebView` surfaces outside the host — see [ADR-0057](decisions/0057-permanent-player-surface-host.md). Enjoy dialogs/sheets ([`enjoy_modal.dart`](../lib/core/theme/widgets/enjoy_modal.dart)) default to the **root** navigator so they paint above the host ([ADR-0065](decisions/0065-enjoy-modals-root-navigator.md)); in-stage chrome uses host `overlayBuilder` instead.
 
 ## Form factor & orientation (ADR-0059)
