+++ test/features/player/application/media_kit_player_engine_linux_test.dart
@@ -1,8 +1,10 @@
 import 'dart:io' show Platform;
 
 import 'package:enjoy_player/features/player/application/player_engine.dart';
+import 'package:enjoy_player/features/player/presentation/widgets/media_kit_video_stage.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:media_kit_video/media_kit_video.dart' show Video;
 
 void main() {
   group('MediaKitPlayerEngine on Linux', () {
@@ -29,36 +31,36 @@ void main() {
     );
 
     testWidgets(
-      'native backend gate lives in buildVideoStage only (issue #658)',
+      'native backend gate lives in the video stage only (issue #658)',
       (tester) async {
         final engine = MediaKitPlayerEngine();
 
-        late BuildContext context;
-        await tester.pumpWidget(
-          Builder(
-            builder: (ctx) {
-              context = ctx;
-              return const SizedBox.shrink();
-            },
+        Future<void> pumpStage() => tester.pumpWidget(
+          MaterialApp(
+            home: Scaffold(
+              body: MediaKitVideoStage(
+                engine: engine,
+                maxWidth: 320,
+                maxHeight: 180,
+              ),
+            ),
           ),
         );
 
-        Widget stage() => engine.buildVideoStage(
-          context: context,
-          maxWidth: 320,
-          maxHeight: 180,
-        );
-
-        // Not allowed yet: a bare placeholder — no Video, hence no
+        // Not allowed yet: a bare placeholder — no [Video], hence no
         // [videoController] and no mpv allocation.
-        expect((stage() as ColoredBox).child, isNull);
+        expect(engine.nativeBackendAllowed, isFalse);
+        await pumpStage();
+        expect(find.byType(Video), findsNothing);
 
         // Engine entry points other than [prepareNativeBackend] must not
         // approve the gate. The old `_player` getter did exactly that (it set
         // `_nativeBackendAllowed = true` while claiming to check it), so the
-        // gate is now stage-only and nothing else can flip it.
+        // gate is stage-only and nothing else can flip it.
         engine.warmVideoSurface();
-        expect((stage() as ColoredBox).child, isNull);
+        expect(engine.nativeBackendAllowed, isFalse);
+        await pumpStage();
+        expect(find.byType(Video), findsNothing);
       },
     );
   });
